qing-py-twilio/requirements.txt
azure-functions
twilio
sendgrid